Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
pymongo==4.6.0